+++ addons/im_livechat/static/src/legacy/widgets/livechat_button.js
@@ -9,9 +9,9 @@ import Widget from 'web.Widget';
 
 import { LIVECHAT_COOKIE_HISTORY } from 'im_livechat.legacy.im_livechat.Constants';
 import Feedback from '@im_livechat/legacy/widgets/feedback/feedback';
-import WebsiteLivechat from '@im_livechat/legacy/models/website_livechat';
-import WebsiteLivechatMessage from '@im_livechat/legacy/models/website_livechat_message';
-import WebsiteLivechatWindow from '@im_livechat/legacy/models/website_livechat_window';
+import PublicLivechat from '@im_livechat/legacy/models/website_livechat';
+import PublicLivechatMessage from '@im_livechat/legacy/models/website_livechat_message';
+import PublicLivechatWindow from '@im_livechat/legacy/models/website_livechat_window';
 
 import { clear } from '@mail/model/model_field_command';
 
@@ -101,7 +101,7 @@ const LivechatButton = Widget.extend({
         options = Object.assign({}, this.options, options, {
             serverURL: this.messaging.livechatButtonView.serverUrl,
         });
-        const message = new WebsiteLivechatMessage(this, data, options);
+        const message = new PublicLivechatMessage(this, data, options);
 
         const hasAlreadyMessage = _.some(this.messaging.livechatButtonView.messages, function (msg) {
             return message.getID() === msg.getID();
@@ -259,7 +259,7 @@ const LivechatButton = Widget.extend({
                 }
             } else {
                 this.messaging.livechatButtonView.update({
-                    livechat: new WebsiteLivechat({
+                    livechat: new PublicLivechat({
                         parent: this,
                         data: livechatData
                     }),
@@ -319,7 +319,7 @@ const LivechatButton = Widget.extend({
             titleColor: this.messaging.livechatButtonView.titleColor,
         };
         this.messaging.livechatButtonView.update({
-            chatWindow: new WebsiteLivechatWindow(this, this.messaging.livechatButtonView.livechat, options),
+            chatWindow: new PublicLivechatWindow(this, this.messaging.livechatButtonView.livechat, options),
         });
         return this.messaging.livechatButtonView.chatWindow.appendTo($('body')).then(() => {
             const cssProps = { bottom: 0 };
